First commit - The end-to-end basics
First version of a C program to generate the data for the FPUT problem,
using the quadratic terms and the Velocity-Verlet method. Furthermore, a
gnuplot script to generate the images files necessary to create an
animation with ffmpeg. All the code and scripts were designed to run on
unix systems with bash, gcc, gnuplot and ffmpeg installed.